requirements.txt
click>=8.0
rich>=13.0.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
cryptography>=42.0.0
pynacl>=1.5.0
python-jose[cryptography]>=3.3.0
ratelimit>=2.2.1
requests>=2.30.0
fastapi>=0.95.0
uvicorn>=0.22.0
gunicorn>=21.2.0
uvloop>=0.17.0
redis>=5.0.0
python-json-logger>=2.0.7
